feat(c,emu): 27 de 81 -- y el paso de un array de arrays estaba MAL
Segunda tanda contra los ficheros de DOOM, con el mismo metodo: compilar los 81
uno a uno y quedarse con el primer error de cada uno.

  antes de hoy:  0 de 81
  primera tanda: 7 de 81
  ahora:        27 de 81

* LAS CAUSAS DE ESTA TANDA, por ficheros que desbloquea cada una

  24  un PARAMETRO que es puntero a funcion -- `boolean (*trav)(intercept_t*)`.
      El declarador se entendia para globales, miembros, locales y typedefs, y
      no aqui: un callback se podia declarar, guardar y llamar, pero NO PASAR.
      `p_map.c` y `p_sight.c` estan construidos sobre pasar el traverser.
  39  `[5][256]` -- solo se leia el primer corchete, asi que el segundo se
      quedaba delante del parser. Son las tablas de gamma de `tables.h` y las
      cajas de los nodos de `doomdata.h`, o sea el formato del mapa en disco.
  20  `int data1, data2, data3, data4;` DENTRO de un agregado. `d_event.h`, el
      evento por el que viaja toda la entrada del juego, mete sus cuatro campos
      en una linea.
   ~  `int alpha, beta;` a nivel de FICHERO. Funcionaba dentro de una funcion y
      no fuera, asi que la diferencia era el AMBITO y no la sintaxis.
   ~  `typedef byte sha1_digest_t[20];` -- un typedef DE UN ARRAY. Una linea de
      `sha1.h`, que entra por `net_defs.h`, que entra por `doomstat.h`: asi
      llego a medio juego.
   ~  `int t[] = { 10, 20, 30 };` -- la lista es la que dice cuanto mide.
   ~  `extern int t[];` -- incompleto, que es justo lo que ese `extern` afirma.
   ~  `inline`, `__inline__`, `__forceinline` -- se consumen y se tiran. Es una
      PETICION y el estandar deja ignorarla; BMO C no inlinea, asi que
      atenderla y no atenderla dan el mismo programa. Lo unico que hacia la
      palabra era parar el fichero.
   ~  Una linea que acaba en `\` NO es una linea (fase 2). Sin eso, una macro
      escrita en varias lineas define un cuerpo de `\` y suelta el resto dentro
      del fichero como si fuera codigo -- y el error cae en una linea que no
      deberia existir.

★★ Y LO QUE SALIO DE DEBAJO: EL PASO DE UN ARRAY DE ARRAYS ERA 8

Aceptar `int grid[2][3]` destapo que **`grid[1][0]` leia `grid[0][2]`**. La
funcion que da el paso de un indice contestaba 8 para cualquier array de
arrays: caia en el caso por defecto. Un paso del indice de fuera es una FILA
ENTERA -- doce bytes ahi, 256 en `gammatable[5][256]`.

Eso compila, corre y **imprime un numero plausible**. Es la clase de fallo que
este banco de pruebas existe para cazar, y lo cazo porque la fila nueva EJECUTA
el programa y compara la salida en vez de mirar bytes.

El paso pasa de `u8` a `u32` en los dos frontends: un `u8` ahi no recorta, DA
LA VUELTA, que es el mismo fallo con una tabla mas grande.

★ Y EL EMULADOR NO TENIA `imul reg, r/m, imm`

Al escalar por doce, el compilador emite `0x6B`. Lo llevaba emitiendo desde
siempre para cualquier paso que no fuera potencia de dos, y **ningun test lo
habia ejecutado nunca**. El emulador hizo lo correcto: panic con el opcode en
la mano en vez de seguir con un valor inventado. Ese panic es el que descubrio
lo de arriba. Anadidos `0x6B` (imm8) y `0x69` (imm32).

* Y UNA REGLA QUE SE ENDURECE

La medida de un array que no se puede calcular al compilar era **un 1 callado**.
Ahora es un error con motivo: un array de un solo hueco compila, corre, y cada
escritura pasada la primera cae encima de la variable de al lado.

* VERIFICACION

306 tests de `bmo-c-front` en verde (297 antes; nueve filas nuevas, una por
causa, y todas ejecutan). Suite completa sin regresiones nuevas -- el 108/110
de `bmo-cpp-front` sigue igual que estaba.

★ Y lo que cierra esto antes de flashear: `build.ps1 -BuildOnly` entero en
verde y **los 27 `.bex` salen BYTE A BYTE IDENTICOS** a los de antes de tocar
el compilador. Ningun programa que hoy corre en el Ryzen cambia.

+++ toolchain/lang/c/src/ast/expr.rs
@@ -52,9 +52,9 @@ pub enum Expr {
     Comma(Vec<Expr>),
     Deref(Box<Expr>),
     AddrOf(Box<Expr>),
-    Subscript(String, Box<Expr>, u8),
+    Subscript(String, Box<Expr>, u32),
     /// arr[i] = val -- antes la asignacion a subscript se DESCARTABA en silencio.
-    AssignSubscript(String, Box<Expr>, u8, Box<Expr>),
+    AssignSubscript(String, Box<Expr>, u32, Box<Expr>),
     /// base[i] donde base es una EXPRESION que da un puntero (p->arr[i],
     /// (a+1)[i]): (base, indice, tipo del elemento). Antes se rechazaba.
     IndexPtr(Box<Expr>, Box<Expr>, TypeSpec),
